+++ transcriptic/cli.py
@@ -606,10 +606,17 @@ def preview(ctx, protocol_name, view):
 @click.pass_context
 @click.option('--tree', '-t', is_flag=True,
               help='Prints a job tree with instructions as leaf nodes')
-# default allowance set to 5 seconds
+# time allowance is on order of seconds
 @click.option('--runtime', "-r", type=click.INT, default=5)
 def summarize(ctx, file, tree, runtime):
-    """Summarize Autoprotocol as a list of plain English steps."""
+    """Summarize Autoprotocol as a list of plain English steps, as well as a
+    visualized job tree contingent upon desired runtime allowance. Example usage 
+    is as follows:
+
+    python my_script.py | transcriptic summarize --tree
+    python my_script.py | transcriptic summarize --tree --runtime 20
+
+    """
     with click.open_file(file, 'r') as f:
         try:
             protocol = json.loads(f.read())